+++ rich/default_styles.py
@@ -60,6 +60,7 @@ DEFAULT_STYLES: Dict[str, Style] = {
     "log.time": Style(color="cyan", dim=True),
     "log.message": Style.null(),
     "log.path": Style(dim=True),
+    "log.thread_name": Style(color="yellow"),
     "repr.ellipsis": Style(color="yellow"),
     "repr.indent": Style(color="green", dim=True),
     "repr.error": Style(color="red", bold=True),